+++ examples/context_layering.rs
@@ -43,15 +43,14 @@ impl GamePlugin {
 
     fn apply_movement(trigger: Trigger<ActionEvent<Move>>, mut players: Query<&mut Transform>) {
         let event = trigger.event();
-        if event.kind.is_fired() {
+        if event.is_fired() {
             let mut transform = players.get_mut(trigger.entity()).unwrap();
             transform.translation += event.value.as_axis3d();
         }
     }
 
     fn rotate(trigger: Trigger<ActionEvent<Rotate>>, mut players: Query<&mut Transform>) {
-        let event = trigger.event();
-        if event.kind.is_started() {
+        if trigger.event().is_started() {
             let mut transform = players.get_mut(trigger.entity()).unwrap();
             transform.rotate_z(FRAC_PI_4);
         }
@@ -62,8 +61,7 @@ impl GamePlugin {
         mut commands: Commands,
         mut players: Query<&mut PlayerColor>,
     ) {
-        let event = trigger.event();
-        if event.kind.is_started() {
+        if trigger.event().is_started() {
             // Change color for visibility.
             let mut color = players.get_mut(trigger.entity()).unwrap();
             color.0 = INDIGO_600.into();
@@ -73,10 +71,8 @@ impl GamePlugin {
     }
 
     fn dive(trigger: Trigger<ActionEvent<Dive>>, mut players: Query<&mut Visibility>) {
-        let event = trigger.event();
-
         // Hide player while diving.
-        let target_visibility = match event.kind {
+        let target_visibility = match **trigger.event() {
             ActionEventKind::Started => Visibility::Hidden,
             ActionEventKind::Completed { .. } => Visibility::Visible,
             _ => return,
@@ -91,8 +87,7 @@ impl GamePlugin {
         mut commands: Commands,
         mut players: Query<&mut PlayerColor>,
     ) {
-        let event = trigger.event();
-        if event.kind.is_fired() {
+        if trigger.event().is_fired() {
             let mut color = players.get_mut(trigger.entity()).unwrap();
             color.0 = Default::default();