+++ examples/context_sharing.rs
@@ -50,15 +50,14 @@ impl GamePlugin {
 
     fn apply_movement(trigger: Trigger<ActionEvent<Move>>, mut players: Query<&mut Transform>) {
         let event = trigger.event();
-        if event.kind.is_fired() {
+        if event.is_fired() {
             let mut transform = players.get_mut(trigger.entity()).unwrap();
             transform.translation += event.value.as_axis3d();
         }
     }
 
     fn rotate(trigger: Trigger<ActionEvent<Rotate>>, mut players: Query<&mut Transform>) {
-        let event = trigger.event();
-        if event.kind.is_started() {
+        if trigger.event().is_started() {
             let mut transform = players.get_mut(trigger.entity()).unwrap();
             transform.rotate_z(FRAC_PI_4);
         }