+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `ActionData::value` to get triggered value from the last update.
 - `ActionData::trigger_events` to trigger events based on the last `ActionData::update`.
 - `BlockBy::events` to block only events. Could be used for chords to avoid triggering required actions.
+- `Deref` for `ActionEvent::kind`.
 - `ContextInstances` to public API and methods to get `ActionData` for an action.
 
 ### Changed